% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_PRESCCO.R
\name{prediction_coverage_rate}
\alias{prediction_coverage_rate}
\title{Empirical prediction coverage}
\usage{
prediction_coverage_rate(
  zeta,
  r,
  test_y_data,
  test_w_data,
  test_delta_data,
  test_z_c_data = NULL,
  test_z_d_data = NULL,
  beta,
  alpha1_star_r,
  tau1,
  w_min,
  w_max
)
}
\arguments{
\item{zeta}{Half-length.}

\item{r}{Residual function, such as \code{r1} or \code{r2}.}

\item{test_y_data, test_w_data, test_delta_data}{Test outcome, observed time,
and event indicator.}

\item{test_z_c_data, test_z_d_data}{Optional test covariates.}

\item{beta}{Fitted outcome-model coefficient vector.}

\item{alpha1_star_r}{Coefficients used by the residual.}

\item{tau1}{Standard deviation used by the residual.}

\item{w_min, w_max}{Truncation bounds.}
}
\value{
The empirical coverage rate.
}
\description{
Evaluates the empirical coverage of a fixed half-length on a test set.
}
\examples{
set.seed(1)
n <- 200
x_vec <- truncnorm::rtruncnorm(n, a = -1, b = 1, mean = 0, sd = 1)
c_vec <- truncnorm::rtruncnorm(n, a = -1, b = 1, mean = 1, sd = 1)
w_vec <- pmin(x_vec, c_vec)
delta_vec <- as.integer(x_vec <= c_vec)
y_vec <- 3 * x_vec + rnorm(n)

beta_fit <- find_beta_cc(
  y_vec, w_vec, delta_vec
)$beta_cc

prediction_coverage_rate(
  zeta = 2,
  r = r1,
  test_y_data = y_vec,
  test_w_data = w_vec,
  test_delta_data = delta_vec,
  beta = beta_fit,
  alpha1_star_r = 0,
  tau1 = 1,
  w_min = -1,
  w_max = 1
)

}
